Refactor Storage tests
